requirements.txt
fastapi==0.111.0

pydantic>=2.0.0
pydantic-ai>=0.0.23

openai>=1.0.0
python-dotenv>=1.0.0

pydantic-settings==2.2.1
typing_extensions>=4.11.0

uvicorn
httpx

py_trello



